Add check for existing database before dropping HolidayFun

+++ 26.2 Section 4-New update.sql
@@ -1,6 +1,8 @@
-DROP DATABASE HolidayFun;
-
+IF DB_ID('HolidayFun') IS NOT NULL
+    DROP DATABASE HolidayFun;
+GO
 CREATE DATABASE HolidayFun;
+GO
 
 USE HolidayFun;
